refactor admin routes and dashboard
