+++ src/components/admin/new-product-form.tsx
@@ -1,0 +1,5 @@
+const NewProductForm = () => {
+  return <div>NewProductForm</div>;
+};
+
+export default NewProductForm;
